Fix TestBackupMain
We'd added code that updated the checksum and filestamp in the
destination xml.
This failed in the test that didn't copy the file

We didn;t used to copy the file as we simply didn't need to. Now we need
the timestamp of the dest file, then it has to exist