Add test case for bug with duplicate files
If we have a file on the src that was a copy of another, then we
mishandled things.
Add a testcase for this
The duplicate detect went wrong because we deleted the duplicate from
the map without a lock between the query, and the modification.
So in the cmd tool, we queried, then deleted, then queried
In the tests we were query, query, delete.